feat(mobile-lan): support SyncClipboard Android routes
Add the official probe and history endpoints used by SyncClipboard clients, document which routes are compatibility shells, and split the long route file by responsibility.
