+++ src-tauri/crates/uc-webserver/Cargo.toml
@@ -16,7 +16,7 @@ uc-daemon-contract = { path = "../uc-daemon-contract" }
 uc-daemon-local = { path = "../uc-daemon-local" }
 uc-observability = { path = "../uc-observability" }
 anyhow = "1.0"
-axum = { version = "0.7", features = ["ws", "tokio", "http1", "http2"] }
+axum = { version = "0.7", features = ["ws", "tokio", "http1", "http2", "multipart"] }
 base64 = "0.22"
 chrono = { version = "0.4", features = ["serde"] }
 futures-util = "0.3"
